Better file upload

+++ hook_system/client/ClientServer/Services/FileService.cs
@@ -47,30 +47,32 @@ namespace ClientServer.Services
 
         // Saves the files in this submission object.
         // The submission needs a valid SubmissionId, as well as the Assignment and Course loaded
-        public async void PersistSubmissionFiles(Submission submission)
+        public void PersistSubmissionFiles(Submission submission)
         {
             var submissionPath = GetSubmissionPath(submission);
             var basePath = Path.Combine(Directory.GetCurrentDirectory(), _rootStorageDirectory, submissionPath);
             submission.FilePath = submissionPath;
                 
             // Save each file in the folder
-            foreach (var formFile in submission.Files)
-            {
-                // Actually save the file, if there is anything to save
-                // Create the filepath for this file
-                var filePath = Path.Combine(basePath, formFile.FileName);
-
-                // Create any necessary folders for this filepath
-                Directory.CreateDirectory(filePath);
-
-                // Remove the folder created that is just the file name
-                Directory.Delete(filePath);
-
-                if (formFile.Length > 0)
+            if (submission.Files != null) {
+                foreach (var formFile in submission.Files)
                 {
-                    using (var stream = new FileStream(filePath, FileMode.Create))
+                    // Actually save the file, if there is anything to save
+                    // Create the filepath for this file
+                    var filePath = Path.Combine(basePath, formFile.FileName);
+
+                    // Create any necessary folders for this filepath
+                    Directory.CreateDirectory(filePath);
+
+                    // Remove the folder created that is just the file name
+                    Directory.Delete(filePath);
+
+                    if (formFile.Length > 0)
                     {
-                        await formFile.CopyToAsync(stream);
+                        using (var stream = new FileStream(filePath, FileMode.Create))
+                        {
+                            formFile.CopyTo(stream);
+                        }
                     }
                 }
             }
